jobs/pacman-awsrules/src/test/java/com/tmobile/cloud/awsrules/ami/AMIEncryptionRuleTest.java
package com.tmobile.cloud.awsrules.ami;

import static org.assertj.core.api.Assertions.assertThatThrownBy;
import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.CoreMatchers.notNullValue;
import static org.junit.Assert.assertThat;
import static org.mockito.Matchers.anyString;
import static org.powermock.api.mockito.PowerMockito.mockStatic;
import static org.powermock.api.mockito.PowerMockito.when;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.HashMap;
import java.util.Map;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.powermock.core.classloader.annotations.PrepareForTest;
import org.powermock.modules.junit4.PowerMockRunner;

import com.tmobile.cloud.awsrules.utils.PacmanUtils;
import com.tmobile.pacman.commons.exception.InvalidInputException;
import com.tmobile.pacman.commons.exception.RuleExecutionFailedExeption;

@RunWith(PowerMockRunner.class)
@PrepareForTest({ PacmanUtils.class})
public class AMIEncryptionRuleTest {

    @InjectMocks
    AMIEncryptionRule amiEncryptionRule;
 
    @Test
    public void executeTest() throws Exception {
        mockStatic(PacmanUtils.class);
        when(PacmanUtils.doesAllHaveValue(anyString(),anyString(),anyString())).thenReturn(true);
        
        when(PacmanUtils.getUnencryptedSnapshotIds(anyString(), anyString())).thenReturn(new ArrayList<>());
        assertThat(amiEncryptionRule.execute(getMapString("r_123 "),getMapString("r_123 ")), is(notNullValue()));
        
        when(PacmanUtils.getUnencryptedSnapshotIds(anyString(), anyString())).thenReturn(Arrays.asList("snapshotabc-1"));
        assertThat(amiEncryptionRule.execute(getMapString("r_123 "),getMapString("r_123 ")), is(notNullValue()));
        
        when(PacmanUtils.getUnencryptedSnapshotIds(anyString(), anyString())).thenThrow(new Exception());
        assertThatThrownBy(() -> amiEncryptionRule.execute(getMapString("r_123 "),getMapString("r_123 "))).isInstanceOf(RuleExecutionFailedExeption.class);
        
        when(PacmanUtils.doesAllHaveValue(anyString(),anyString(),anyString())).thenReturn(false);
        assertThatThrownBy(
                () -> amiEncryptionRule.execute(getMapString("r_123 "),getMapString("r_123 "))).isInstanceOf(InvalidInputException.class);
        
    }
    
    @Test
    public void getHelpTextTest(){
        assertThat(amiEncryptionRule.getHelpText(), is(notNullValue()));
    }
    
    public static Map<String, String> getMapString(String passRuleResourceId) {
        Map<String, String> commonMap = new HashMap<>();
        commonMap.put("executionId", "1234");
        commonMap.put("_resourceid", passRuleResourceId);
        commonMap.put("severity", "high");
        commonMap.put("ruleCategory", "security");
        commonMap.put("accountid", "12345");
        commonMap.put("ruleId", "Aws_AMI_encryption_version-1_aws_enable_encryption_ami");
        commonMap.put("policyId", "Aws_AMI_encryption_version-1");
        commonMap.put("policyVersion", "version-1");
        commonMap.put("esAmiBlockDeviceMappingUrl", "esFindingsUrl");
        commonMap.put("snapshotid", "snapshotabc-1");
        return commonMap;
    }
  
    
}
